fix: 修复 MSVC 构建的两处编译错误
fix: fix two MSVC build errors

- DasStringImpl.cpp: u_strCompare 的 length 形参为 int32_t，直接传入
  size_t 触发 C4267（警告被视为错误）；按同文件既有写法补 static_cast<int32_t>。
- DasAddIdlExport.cmake: Node export target 缺少 /bigobj，autogen 的
  node_export.cpp 节数超限触发 C1128；与同文件 Lua export 一致补上 MSVC /bigobj。

- DasStringImpl.cpp: u_strCompare length params are int32_t; passing size_t
  triggered C4267 (warning treated as error). Add static_cast<int32_t> matching
  the existing pattern in the same file.
- DasAddIdlExport.cmake: Node export target lacked /bigobj; the autogenerated
  node_export.cpp exceeded the COFF section limit (C1128). Add MSVC /bigobj,
  consistent with the Lua export target in the same file.

～(∠・ω< )⌒★ GLM-5.2 代表 Dusk 创建了这个提交

+++ das/Core/ForeignInterfaceHost/src/DasStringImpl.cpp
@@ -348,7 +348,13 @@ DasBool DasStringCppImpl::Equals(IDasReadOnlyString* other) noexcept
     {
         return false;
     }
-    return u_strCompare(lhs_utf16, lhs_size, rhs_utf16, rhs_size, false) == 0;
+    return u_strCompare(
+               lhs_utf16,
+               static_cast<int32_t>(lhs_size),
+               rhs_utf16,
+               static_cast<int32_t>(rhs_size),
+               false)
+           == 0;
 }
 
 /**
